Begin building the layout for the first question of the quiz
It might be nice to have the questions on individual cards which slide
in and out like Reigns using material design elements, but I think the
best thing to do for now is have them all one one page. Here I have
begun to build the layout structure, adding the first question and
answer options. I will pretty much just copy/paste this a few times
now and make modifications as needed to fill out the rest of the quiz
questions.