Complete a rough layout including RadioButton, CheckBox, and EditText
I have built the basic framework for the app. I wish I had found a
more modular way of doing it, because it involved a lot of copy and
pasting and changing one character per line which felt pretty cheap.
Also, I can't easily move questions around in order, because they are
named for their number. Perhaps a better way of doing this would have
been to use an array or something, or just build the objects in Java.

Next I have to copy the the questions and answers from my paper notes,
then write the java code.