Enter each answer into strings.xml and fix string references.
There were about four answers for each question, but the questions were
shorter than the answers so overall it was about the same amount of
typing.

Next I need to make the code to handle the form submission, and I should
edit the styles to make sure buttons align and stuff. I kind of want to
do that first but I should probably do the coding/scoring next.